Improvements :make the code more readable for and make the others compenet
